+++ changelog.md
@@ -59,3 +59,4 @@
 * assume current year with input `spacetime('July 5th')`
 * do larger `.diff()` operations (gt 1yr) in fast-mode
 * support BC formatting of input and in `s.format('year')`
+* bug fix for iso-format hour-offsets (#58)